use client for header and footer links to function
