+++ src/app/api/v1/route.ts
@@ -16,3 +16,5 @@ export async function GET() {
 
   return NextResponse.json(data);
 }
+
+export async function POST() {}
