+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import "./globals.css";
-import Link from "next/link";
 import { Fira_Code } from "next/font/google";
-import { Footer } from "../components";
+import { Footer, Header } from "../components";
 
 const firaCode = Fira_Code({
   subsets: ["latin"],
@@ -22,19 +21,7 @@ export default async function RootLayout({
       <body
         className={`${firaCode.className} flex min-h-screen items-center justify-center bg-zinc-950 text-zinc-200`}
       >
-        <nav className="fixed top-0 flex w-full items-center justify-between bg-inherit px-8 py-6 text-lg text-zinc-200">
-          <Link href="/" className="flex items-center gap-4">
-            <span className="text-2xl">css2wind</span>
-          </Link>
-          <div className="hidden items-center justify-center gap-10 md:flex">
-            <a
-              className="flex items-center gap-4 px-4 py-2 hover:outline hover:outline-zinc-100"
-              href="/faq"
-            >
-              Learn more about TailwindCSS
-            </a>
-          </div>
-        </nav>
+        <Header />
         {children}
         <Footer />
       </body>
