+++ src/components/index.ts
@@ -1,1 +1,3 @@
 export { default as Footer } from "./Footer";
+export { default as Tutorial } from "./Tutorial";
+export { default as Header } from "./Header";
